Adding working Project 1 to GitHub.

+++ prj/1/Driver_prj1.java
@@ -3,7 +3,7 @@
 // Course: CMPT220
 // Assignment: Making a Convolution (Project 1)
 // Due date: March 30th, 2017
-// Version 1.3
+// Version 1.5
 
 import java.util.Scanner;
 
@@ -25,11 +25,11 @@ public class Driver_prj1 {
     double[] vector = convolveVectors(vFirst, vSecond); // Where a double/int conversion issue occurs
     
     for (int i = 0; i < vector.length; i++) // Doesn't work due to issue lised above
-      System.out.print(Math.round(vector[i]) + " ");
+      System.out.print(Math.round(vector[i]) + " "); // prints out the answers to project 1
   }
 
-  public static int[] convolveVectors(int[] vFirst, int[] vSecond) {
-    int[] vResult = new int[(vFirst.length + vSecond.length) - 1];
+  public static double[] convolveVectors(double[] vFirst, double[] vSecond) {
+    double[] vResult = new double[(vFirst.length + vSecond.length) - 1];
     for (int idx = 0; idx < vResult.length; idx++){ // checks to see if index is less than vResult length
       for (int shift = 0; shift < vSecond.length; shift++){ // checks to see if the shift goes through all values of vSecond
         if (((idx - shift) >= 0) && ((idx - shift) < vFirst.length)) // functions is here to avoid out of bounds issues
